lib: rename modules js source to lower snake_case
This commit renames all JavaScript source files in lib/internal/modules
to lower snake_case.
