+++ lib/internal/modules/cjs/loader.js
@@ -54,8 +54,9 @@ module.exports = Module;
 
 // these are below module.exports for the circular reference
 const asyncESM = require('internal/process/esm_loader');
-const ModuleJob = require('internal/modules/esm/ModuleJob');
-const createDynamicModule = require('internal/modules/esm/CreateDynamicModule');
+const ModuleJob = require('internal/modules/esm/module_job');
+const createDynamicModule = require(
+  'internal/modules/esm/create_dynamic_module');
 const {
   CHAR_UPPERCASE_A,
   CHAR_LOWERCASE_A,
