+++ lib/internal/modules/esm/loader.js
@@ -6,11 +6,12 @@ const {
   ERR_MISSING_DYNAMIC_INTSTANTIATE_HOOK,
   ERR_UNKNOWN_MODULE_FORMAT
 } = require('internal/errors').codes;
-const ModuleMap = require('internal/modules/esm/ModuleMap');
-const ModuleJob = require('internal/modules/esm/ModuleJob');
-const defaultResolve = require('internal/modules/esm/DefaultResolve');
-const createDynamicModule = require('internal/modules/esm/CreateDynamicModule');
-const translators = require('internal/modules/esm/Translators');
+const ModuleMap = require('internal/modules/esm/module_map');
+const ModuleJob = require('internal/modules/esm/module_job');
+const defaultResolve = require('internal/modules/esm/default_resolve');
+const createDynamicModule = require(
+  'internal/modules/esm/create_dynamic_module');
+const translators = require('internal/modules/esm/translators');
 
 const FunctionBind = Function.call.bind(Function.prototype.bind);
