+++ lib/internal/process/esm_loader.js
@@ -7,7 +7,7 @@ const {
 } = internalBinding('module_wrap');
 
 const { getURLFromFilePath } = require('internal/url');
-const Loader = require('internal/modules/esm/Loader');
+const Loader = require('internal/modules/esm/loader');
 const path = require('path');
 const { URL } = require('url');
