+++ node.gyp
@@ -106,12 +106,12 @@
       'lib/internal/linkedlist.js',
       'lib/internal/modules/cjs/helpers.js',
       'lib/internal/modules/cjs/loader.js',
-      'lib/internal/modules/esm/Loader.js',
-      'lib/internal/modules/esm/CreateDynamicModule.js',
-      'lib/internal/modules/esm/DefaultResolve.js',
-      'lib/internal/modules/esm/ModuleJob.js',
-      'lib/internal/modules/esm/ModuleMap.js',
-      'lib/internal/modules/esm/Translators.js',
+      'lib/internal/modules/esm/loader.js',
+      'lib/internal/modules/esm/create_dynamic_module.js',
+      'lib/internal/modules/esm/default_resolve.js',
+      'lib/internal/modules/esm/module_job.js',
+      'lib/internal/modules/esm/module_map.js',
+      'lib/internal/modules/esm/translators.js',
       'lib/internal/safe_globals.js',
       'lib/internal/net.js',
       'lib/internal/os.js',
