+++ test/message/esm_display_syntax_error_import.out
@@ -3,4 +3,4 @@ file:///*/test/message/esm_display_syntax_error_import.mjs:6
   notfound
   ^^^^^^^^
 SyntaxError: The requested module '../fixtures/es-module-loaders/module-named-exports' does not provide an export named 'notfound'
-    at ModuleJob._instantiate (internal/modules/esm/ModuleJob.js:*:*)
+    at ModuleJob._instantiate (internal/modules/esm/module_job.js:*:*)
