+++ test/message/esm_display_syntax_error_import_module.out
@@ -3,4 +3,4 @@ file:///*/test/fixtures/es-module-loaders/syntax-error-import.mjs:1
 import { foo, notfound } from './module-named-exports';
               ^^^^^^^^
 SyntaxError: The requested module './module-named-exports' does not provide an export named 'notfound'
-    at ModuleJob._instantiate (internal/modules/esm/ModuleJob.js:*:*)
+    at ModuleJob._instantiate (internal/modules/esm/module_job.js:*:*)
